back/requirements.txt
tqdm~=4.67.1
evaluate~=0.4.0
datasets~=2.21.0
gliner~=0.2.15
neo4j~=5.28.1
langchain-core~=0.3.45
langchain-openai~=0.3.4
pydantic~=2.10.6
spacy~=3.8.3
glirel~=1.0.1
langchain-community~=0.3.19
numpy~=1.26.3
pandas~=2.2.3
torch==2.6.0+cu126
matplotlib~=3.10.1
seaborn~=0.13.2
transformers~=4.49.0
scikit-learn~=1.6.1
setuptools~=75.8.0
jsonlines~=4.0.0
openai~=1.65.5
langchain_huggingface~=0.1.2
faiss-cpu~=1.10.0